backend/pyproject.toml
[project]
name = "logogen"
version = "0.1.0"
description = "Local brand design system generator — MCP server for LLM-based marketing teams"
readme = "README.md"
authors = [
    { name = "Dave Barton", email = "8168910+gofastercloud@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic==2.11.3",
    "python-dotenv==1.1.0",
    "mlx-lm==0.24.1",
    "mflux==0.13.1",
    "aiosqlite==0.21.0",
    "Pillow>=10.4.0",
    "drawsvg==2.4.0",
    "vtracer==0.6.10",
    "mcp[cli]==1.9.4",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
]

[build-system]
requires = ["uv_build>=0.11.0,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests that load real models (deselect with '-m \"not slow\"')",
]
